+++ src/renderers/_base.ts
@@ -216,6 +216,60 @@ const edges2corners = new Map<Orientation, IEdge[]>([
 ]);
 
 /**
+ * For the generic "pieces" area
+ */
+export interface IPiecesArea {
+    type: "pieces";
+    pieces: [string, ...string[]];
+    label: string;
+}
+
+/** Helper functions for drawing edge click handlers */
+const sortPoints = (a: [number,number], b: [number,number]) => {
+    if (a[0] === b[0]) {
+        if (a[1] === b[1]) {
+            return 0;
+        } else {
+            return a[1] - b[1];
+        }
+    } else {
+        return a[0] - b[0];
+    }
+};
+const pts2id = (a: [number,number], b: [number,number]): string => {
+    const x = a.map(n => Math.trunc(n * 1000) / 1000) as [number,number];
+    const y = b.map(n => Math.trunc(n * 1000) / 1000) as [number,number];
+    return [x,y].sort(sortPoints).map(p => p.join(",")).join(" ");
+}
+type CompassDirection = "N"|"NE"|"E"|"SE"|"S"|"SW"|"W"|"NW";
+interface IEdge {
+    dir: CompassDirection;
+    corners: [0|1|2|3|4|5,0|1|2|3|4|5];
+}
+const oppDir = new Map<CompassDirection,CompassDirection>([
+    ["N","S"],["NE","SW"],["E","W"],["SE","NW"],
+    ["S","N"],["SW","NE"],["W","E"],["NW","SE"],
+]);
+const edges2corners = new Map<Orientation, IEdge[]>([
+    [Orientation.FLAT, [
+        {dir: "N", corners: [5,0]},
+        {dir: "NE", corners: [0,1]},
+        {dir: "SE", corners: [1,2]},
+        {dir: "S", corners: [2,3]},
+        {dir: "SW", corners: [3,4]},
+        {dir: "NW", corners: [4,5]},
+    ]],
+    [Orientation.POINTY, [
+        {dir: "NE", corners: [5,0]},
+        {dir: "E", corners: [0,1]},
+        {dir: "SE", corners: [1,2]},
+        {dir: "SW", corners: [2,3]},
+        {dir: "W", corners: [3,4]},
+        {dir: "NW", corners: [4,5]},
+    ]],
+]);
+
+/**
  * An infinite generator for creating column labels from an initial string of characters.
  * With the English alphabet, you would get a-z, then aa-az-ba-zz, then aaa etc.
  *
@@ -1658,7 +1712,7 @@ export abstract class RendererBase {
                         continue;
                     }
                     seenEdges.add(vid);
-                    const edgeLine = board.line(x1, y1, x2, y2).stroke({ width: baseStroke, color: baseColour, opacity: baseOpacity }).translate(x,y);
+                    const edgeLine = board.line(x1, y1, x2, y2).stroke({ width: baseStroke, color: baseColour, opacity: baseOpacity, linecap: "round" }).translate(x,y);
                     if (this.options.rotate === 180) {
                         edgeLine.click(() => this.options.boardClick!(height - hex.row - 1, width - hex.col - 1, oppDir.get(edge.dir)!));
                     } else {
@@ -2616,7 +2670,7 @@ export abstract class RendererBase {
                                 yTo = south.y;
                                 break;
                         }
-                        svgGroup.line(xFrom, yFrom, xTo, yTo).stroke({width: baseStroke * multiplier, color: colour});
+                        svgGroup.line(xFrom, yFrom, xTo, yTo).stroke({width: baseStroke * multiplier, color: colour, linecap: "round"});
                     } else if ( (hexGrid !== undefined) && (hexWidth !== undefined) && (hexHeight !== undefined) && ( (style.startsWith("hex-odd")) || (style.startsWith("hex-even")) ) ) {
                         let row = marker.cell.row as number;
                         let col = marker.cell.col as number;
@@ -2636,7 +2690,7 @@ export abstract class RendererBase {
                                 const [idx1, idx2] = edge.corners;
                                 const {x: xFrom, y: yFrom} = hex.corners[idx1];
                                 const {x: xTo, y: yTo} = hex.corners[idx2];
-                                svgGroup.line(xFrom, yFrom, xTo, yTo).stroke({width: baseStroke * multiplier, color: colour});
+                                svgGroup.line(xFrom, yFrom, xTo, yTo).stroke({width: baseStroke * multiplier, color: colour, linecap: "round"});
                             }
                         }
                     }
@@ -3034,5 +3088,112 @@ export abstract class RendererBase {
         // set the viewbox and return
         nested.viewbox(0, 0, (height * 1.1) + maxScaledWidth, (height * key.list.length) + (buffer * (key.list.length - 1)));
         return nested;
+    }
+
+    /**
+     * For placing a generic `pieces` area at the bottom of the board.
+     *
+     * @protected
+     * @param {GridPoints} gridPoints
+     * @memberof RendererBase
+     * @returns Svg
+     */
+    protected piecesArea(gridPoints: GridPoints) {
+        if (this.rootSvg === undefined) {
+            throw new Error("Can't place a `pieces` area until the root SVG is initialized!");
+        }
+        if ( (this.json !== undefined) && (this.json.areas !== undefined) && (Array.isArray(this.json.areas)) && (this.json.areas.length > 0) ) {
+            const areas = this.json.areas.filter((x) => x.type === "pieces") as IPiecesArea[];
+            const boardBottom = gridPoints[gridPoints.length - 1][0].y + this.cellsize;
+            // Width in number of cells, taking the maximum board width
+            const boardWidth = Math.max(...gridPoints.map(r => r.length));
+            let placeY = boardBottom + (this.cellsize / 2);
+            for (let iArea = 0; iArea < areas.length; iArea++) {
+                const area = areas[iArea];
+                const numPieces = area.pieces.length;
+                const numRows = Math.ceil(numPieces / boardWidth);
+                const textHeight = 10; // the allowance for the label
+                const cellsize = this.cellsize * 0.75;
+                const areaWidth = cellsize * boardWidth;
+                const areaHeight = (textHeight * 2) + (cellsize * numRows);
+                let markWidth = 0;
+                let markColour: string|undefined;
+                if ( ("ownerMark" in area) && (area.ownerMark !== undefined) ) {
+                    markWidth = 15;
+                    if (typeof area.ownerMark === "number") {
+                        markColour = this.options.colours[area.ownerMark - 1];
+                    } else {
+                        markColour = area.ownerMark as string;
+                    }
+                }
+                const nested = this.rootSvg.nested().id(`_pieces${iArea}`).size(areaWidth+2, areaHeight+2).viewbox(-1 - markWidth, -1, areaWidth+2+markWidth, areaHeight+2);
+                if ("background" in area) {
+                    nested.rect(areaWidth,areaHeight).fill(area.background as string);
+                }
+                for (let iPiece = 0; iPiece < area.pieces.length; iPiece++) {
+                    const used: [SVGUse, number][] = [];
+                    const p = area.pieces[iPiece];
+                    const row = Math.floor(iPiece / boardWidth);
+                    const col = iPiece % boardWidth;
+                    const piece = this.rootSvg.findOne("#" + p) as Svg;
+                    if ( (piece === null) || (piece === undefined) ) {
+                        throw new Error(`Could not find the requested piece (${p}). Each piece in the stack *must* exist in the \`legend\`.`);
+                    }
+                    let sheetCellSize = piece.viewbox().h;
+                    if ( (sheetCellSize === null) || (sheetCellSize === undefined) ) {
+                        sheetCellSize = piece.attr("data-cellsize") as number;
+                        if ( (sheetCellSize === null) || (sheetCellSize === undefined) ) {
+                            throw new Error(`The glyph you requested (${p}) does not contain the necessary information for scaling. Please use a different sheet or contact the administrator.`);
+                        }
+                    }
+                    const use = nested.use(piece);
+                    if (this.options.boardClick !== undefined) {
+                        use.click((e: Event) => {this.options.boardClick!(-1, -1, p); e.stopPropagation();});
+                    }
+                    used.push([use, piece.viewbox().h]);
+                    const factor = (cellsize / sheetCellSize);
+                    const newx = col * cellsize;
+                    const newy = (textHeight * 2) + (row * cellsize);
+                    use.dmove(newx, newy);
+                    use.scale(factor, newx, newy);
+                }
+
+                // add marker line if indicated
+                if ( (markWidth > 0) && (markColour !== undefined) ) {
+                    nested.line(markWidth * -1, 0, markWidth * -1, nested.bbox().height).stroke({width: markWidth, color: markColour});
+                }
+
+                // Add area label
+                const tmptxt = this.rootSvg.text(area.label).font({size: textHeight, anchor: "start", fill: "#000"});
+                const txtWidth = tmptxt.bbox().w;
+                tmptxt.remove();
+                nested.width(Math.max(areaWidth, txtWidth));
+                const txt = nested.text(area.label);
+                txt.font({size: textHeight, anchor: "start", fill: "#000"})
+                    .attr("alignment-baseline", "hanging")
+                    .attr("dominant-baseline", "hanging")
+                    .move(0, 0);
+
+                // Now place the whole group below the board
+                // const placed = this.rootSvg.use(nested);
+                nested.move(gridPoints[0][0].x, placeY);
+                placeY += nested.bbox().height + (this.cellsize * 0.5);
+            }
+        }
+    }
+
+    protected backFill() {
+        if (this.rootSvg === undefined) {
+            throw new Error("Can't add a back fill unless the root SVG is initialized!");
+        }
+        if ( (this.json === undefined) || (this.json.board === undefined) ) {
+            throw new Error("Can't add a back fill unless the JSON is initialized!");
+        }
+        if (this.json.board !== null) {
+            if ( ("backFill" in this.json.board) && (this.json.board.backFill !== undefined) && (this.json.board.backFill !== null) ) {
+                const bbox = this.rootSvg.bbox();
+                this.rootSvg.rect(bbox.width + 20, bbox.height + 20).move(bbox.x - 10, bbox.y - 10).fill(this.json.board.backFill).back();
+            }
+        }
     }
 }